jp_en_translator_project/jp_en_translator_project/requirements.txt
torch
transformers
TTS
openai-whisper
sounddevice
sacrebleu
datasets
streamlit